+++ postgres/check.py
@@ -505,7 +505,7 @@ SELECT s.schemaname,
                 self.log.warn('Failed to parse config element=%s, check syntax' % str(element))
         return config
 
-    def _collect_stats(self, key, db, instance_tags, relations, custom_metrics, function_metrics, count_metrics, database_size_metrics, do_collect_default_db, interface_error, programming_error):
+    def _collect_stats(self, key, db, instance_tags, relations, custom_metrics, function_metrics, count_metrics, database_size_metrics, collect_default_db, interface_error, programming_error):
         """Query pg_stat_* for various metrics
         If relations is not an empty list, gather per-relation metrics
         on top of that.
@@ -532,8 +532,8 @@ SELECT s.schemaname,
             # FIXME: constants shouldn't be modified
             self.DB_METRICS['metrics'] = db_instance_metrics
             self.DB_METRICS_WITH_DEFAULT['metrics'] = db_instance_metrics
-            # DB_METRICS query is chosen according to do_collect_default_db parameter
-            if not(do_collect_default_db):
+            # DB_METRICS query is chosen according to collect_default_db parameter
+            if not(collect_default_db):
                 metric_scope.append(self.DB_METRICS)
             else:
                 metric_scope.append(self.DB_METRICS_WITH_DEFAULT)
@@ -752,7 +752,7 @@ SELECT s.schemaname,
         # Default value for `count_metrics` is True for backward compatibility
         count_metrics = _is_affirmative(instance.get('collect_count_metrics', True))
         database_size_metrics = _is_affirmative(instance.get('collect_database_size_metrics', True))
-        do_collect_default_db = _is_affirmative(instance.get('do_collect_default_database', False))
+        collect_default_db = _is_affirmative(instance.get('collect_default_database', False))
 
         if relations and not dbname:
             self.warning('"dbname" parameter must be set when using the "relations" parameter.')
@@ -787,11 +787,11 @@ SELECT s.schemaname,
             db = self.get_connection(key, host, port, user, password, dbname, ssl, connect_fct)
             version = self._get_version(key, db)
             self.log.debug("Running check against version %s" % version)
-            self._collect_stats(key, db, tags, relations, custom_metrics, function_metrics, count_metrics, database_size_metrics, do_collect_default_db, interface_error, programming_error)
+            self._collect_stats(key, db, tags, relations, custom_metrics, function_metrics, count_metrics, database_size_metrics, collect_default_db, interface_error, programming_error)
         except ShouldRestartException:
             self.log.info("Resetting the connection")
             db = self.get_connection(key, host, port, user, password, dbname, ssl, connect_fct, use_cached=False)
-            self._collect_stats(key, db, tags, relations, custom_metrics, function_metrics, count_metrics, database_size_metrics, do_collect_default_db, interface_error, programming_error)
+            self._collect_stats(key, db, tags, relations, custom_metrics, function_metrics, count_metrics, database_size_metrics, collect_default_db, interface_error, programming_error)
 
         if db is not None:
             service_check_tags = self._get_service_check_tags(host, port, dbname)